+++ test/utils/utils.js
@@ -16,13 +16,13 @@
 
 // To be used as a tag
 exports.sane = function() {
-  var str = String.raw.apply(String, arguments);
+  let str = String.raw.apply(String, arguments);
 
-  var match = str.match(/^\n( *)/m);
+  let match = str.match(/^\n( *)/m);
   if (!match) throw new Error('sane string must start with a \\n is:' + str);
-  var spaces = match[1].length;
+  let spaces = match[1].length;
 
-  var lines = str.split('\n');
+  let lines = str.split('\n');
   lines.shift(); // Remove the first empty lines
   lines = lines.map((line) => line.substr(spaces)); // Remove indentation
   if (lines[lines.length - 1] === '') lines.pop(); // Remove last line if empty